# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Smart Blood & Emergency Donor Network API
  version: 1.0.0
  description: Recommended hackathon API contract for the smart blood donor network.
servers:
  - url: http://localhost:8000
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: {type: integer}
...
  schemas:
    RegisterRequest:
      type: object
      required: [name, email, phone, password, role]
      properties:
        name: {type: string, example: Ali Ahmed}
        email: {type: string, format: email, example: ali@example.com}
        phone: {type: string, example: "+923001234567"}
        password: {type: string, format: password}
        role:
          type: string
          enum: [ADMIN, REQUESTER, DONOR, HOSPITAL]
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: {type: string, format: email}
        password: {type: string, format: password}
...
    BloodRequestCreate:
      type: object
      required: [hospital_id, blood_group, units_required, required_before, description, latitude, longitude]
      properties:
        hospital_id: {type: integer}
        blood_group: {type: string, example: B+}
        units_required: {type: integer, minimum: 1, example: 2}
        required_before: {type: string, format: date-time}
        description: {type: string}
        latitude: {type: number, format: double}
        longitude: {type: number, format: double}
...
paths:
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RegisterRequest'}
      responses:
        '201': {description: Registered successfully}
        '400': {description: Invalid request}
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LoginRequest'}
      responses:
        '200': {description: Login successful}
        '401': {description: Invalid credentials}
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get current user
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Current user}
...
  /api/requests:
    get:
      tags: [Blood Requests]
      summary: List blood requests
      security: [{bearerAuth: []}]
      parameters:
        - {name: status, in: query, schema: {type: string}}
        - {name: blood_group, in: query, schema: {type: string}}
        - {name: urgency, in: query, schema: {type: string}}
        - {name: page, in: query, schema: {type: integer}}
        - {name: page_size, in: query, schema: {type: integer}}
      responses:
        '200': {description: Request list}
    post:
      tags: [Blood Requests]
      summary: Create blood request
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/BloodRequestCreate'}
      responses:
        '201': {description: Blood request created}
  /api/requests/{id}:
    get:
      tags: [Blood Requests]
      summary: Get blood request
      security: [{bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses:
        '200': {description: Request details}
    patch:
      tags: [Blood Requests]
      summary: Update blood request
      security: [{bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/Id'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object}
      responses:
        '200': {description: Request updated}
...
  /api/dashboard/requests:
    get:
      tags: [Dashboard]
      summary: Request analytics
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Request analytics}
...